Refactor add_custom_log_attributes logic into new class
