+++ lib/new_relic/agent/log_event_aggregator.rb
@@ -4,6 +4,7 @@
 
 require 'new_relic/agent/event_aggregator'
 require 'new_relic/agent/log_priority'
+require 'new_relic/agent/log_event_attributes'
 
 module NewRelic
   module Agent
@@ -24,9 +25,6 @@ module NewRelic
       FORWARDING_SUPPORTABILITY_FORMAT = 'Supportability/Logging/Forwarding/Ruby/%s'.freeze
       DECORATING_SUPPORTABILITY_FORMAT = 'Supportability/Logging/LocalDecorating/Ruby/%s'.freeze
       MAX_BYTES = 32768 # 32 * 1024 bytes (32 kibibytes)
-      MAX_ATTRIBUTE_COUNT = 240 # limit is 255, assume we send 15 attributes by default
-      ATTRIBUTE_KEY_CHARACTER_LIMIT = 255
-      ATTRIBUTE_VALUE_CHARACTER_LIMIT = 4094
 
       named :LogEventAggregator
       buffer_class PrioritySampledBuffer
@@ -41,7 +39,7 @@ module NewRelic
       DECORATING_ENABLED_KEY = :'application_logging.local_decorating.enabled'
       LOG_LEVEL_KEY = :'application_logging.forwarding.log_level'
 
-      attr_reader :custom_attributes, :already_warned_custom_attribute_count_limit
+      attr_reader :attributes
 
       def initialize(events)
         super(events)
@@ -50,8 +48,7 @@ module NewRelic
         @seen_by_severity = Hash.new(0)
         @high_security = NewRelic::Agent.config[:high_security]
         @instrumentation_logger_enabled = NewRelic::Agent::Instrumentation::Logger.enabled?
-        @custom_attributes = {}
-        @already_warned_custom_attribute_count_limit = false
+        @attributes = NewRelic::Agent::LogEventAttributes.new
         register_for_done_configuring(events)
       end
 
@@ -123,18 +120,8 @@ module NewRelic
         ]
       end
 
-      def add_custom_attributes(attributes)
-        return if already_warned_custom_attribute_count_limit
-
-        attributes.each do |key, value|
-          next if absent?(key) || absent?(value)
-
-          add_custom_attribute(key, value)
-        end
-      end
-
-      def absent?(value)
-        value.nil? || (value.respond_to?(:empty?) && value.empty?)
+      def add_custom_attributes(custom_attributes)
+        attributes.add_custom_attributes(custom_attributes)
       end
 
       # Because our transmission format (MELT) is different than historical
@@ -153,7 +140,7 @@ module NewRelic
         # sent by classic logs-in-context
         common_attributes.delete(ENTITY_TYPE_KEY)
 
-        common_attributes.merge!(NewRelic::Agent.agent.log_event_aggregator.custom_attributes)
+        common_attributes.merge!(NewRelic::Agent.agent.log_event_aggregator.attributes.custom_attributes)
 
         _, items = data
         payload = [{
@@ -175,8 +162,7 @@ module NewRelic
           @seen_by_severity.clear
         end
 
-        @custom_attributes = {}
-        @already_warned_custom_attribute_count_limit = false
+        attributes.reset!
         super
       end
 
@@ -185,64 +171,6 @@ module NewRelic
       end
 
       private
-
-      def add_custom_attribute(key, value)
-        if custom_attributes.size >= MAX_ATTRIBUTE_COUNT
-          NewRelic::Agent.logger.warn(
-            'Too many custom log attributes defined. ' \
-            "Only taking the first #{MAX_ATTRIBUTE_COUNT}."
-          )
-          @already_warned_custom_attribute_count_limit = true
-          return
-        end
-
-        @custom_attributes.merge!(truncate_attributes(key_to_string(key), value))
-      end
-
-      def key_to_string(key)
-        key.is_a?(String) ? key : key.to_s
-      end
-
-      class TruncationError < StandardError
-        attr_reader :attribute, :limit
-
-        def initialize(attribute, limit, msg = "Can't truncate")
-          @attribute = attribute
-          @limit = limit
-          super(msg)
-        end
-      end
-
-      def truncate_attributes(key, value)
-        key = truncate_attribute(key, ATTRIBUTE_KEY_CHARACTER_LIMIT)
-        value = truncate_attribute(value, ATTRIBUTE_VALUE_CHARACTER_LIMIT)
-        {key => value}
-      rescue TruncationError => e
-        NewRelic::Agent.logger.warn(
-          "Dropping custom log attribute #{key} => #{value} \n" \
-          "Length exceeds character limit of #{e.limit}. Can't truncate: #{e.attribute}"
-        )
-        {}
-      end
-
-      def truncate_attribute(attribute, limit)
-        case attribute
-        when Integer
-          if attribute.digits.length > limit
-            raise TruncationError.new(attribute, limit)
-          end
-        when Float
-          if attribute.to_s.length > limit
-            raise TruncationError.new(attribute, limit)
-          end
-        when String, Symbol
-          if attribute.length > limit
-            attribute = attribute.slice(0..(limit - 1))
-          end
-        end
-
-        attribute
-      end
 
       # We record once-per-connect metrics for enabled/disabled state at the
       # point we consider the configuration stable (i.e. once we've gotten SSC)
